+++ app/aktive.py
@@ -13,6 +13,7 @@ from drafthorse.models.accounting import ApplicableTradeTax as DH_ApplicableTrad
 from drafthorse.models.document import Document as DH_Document
 from drafthorse.models.note import IncludedNote as DH_IncludedNote
 from drafthorse.models.party import TaxRegistration as DH_TaxRegistration
+from drafthorse.models.payment import PaymentTerms as DH_PaymentTerms
 from drafthorse.models.tradelines import LineItem as DH_LineItem
 
 
@@ -310,105 +311,170 @@ class Abrechnung:
             out += " "+str(self.getprojectdate())
         return out
 
-    def factur_x(self):
+    def factur_x(self) -> bytes:
+        """
+        Erstellt aus der Abrechnung eine E-Rechnung,
+        die dem Standard Factur-X Extended (auch bekannt als
+        ZUGFeRD Extended) entspricht. Diese wird als binäre
+        XML-Datei zurückgegeben.
+        """
 
+        # General information
         doc = DH_Document()
         doc.context.guideline_parameter.id = "urn:cen.eu:en16931:2017#conformant#urn:factur-x.eu:1p0:extended"
         doc.header.id = "AKTIVE"+datetime.now().strftime("%Y%m%d%H%M%S")
         doc.header.name = self._NAME.upper()
-        doc.header.type_code = "389" # Self billed invoice
         doc.header.issue_date_time = date.today()
         doc.header.languages.add("de")
 
         if self.getprojectname():
             note = DH_IncludedNote()
-            note.content_code = 'Project'
-            note.content.add(self.getprojectname())
+            note.content_code = 'PROJECT'
+            datestring = format_date(self.getprojectdate(),
+                                     format="long",locale="de_DE")
+            note.content.add(self.getprojectname()+" ("+datestring+")")
+            note.subject_code = "ACD" # Reason
             doc.header.notes.add(note)
 
-        doc.trade.agreement.seller.name = CONTACT['Name']
-        doc.trade.agreement.seller.address.line_one = CONTACT['LineOne']
+        # Determine seller and buyer
+        if self.gettotal() > 0:
+            mode = 1
+            doc.header.type_code = "389" # Self billed invoice
+            adfc = doc.trade.agreement.seller
+            user = doc.trade.agreement.buyer
+        else:
+            mode = -1
+            doc.header.type_code = "380" # Commercial invoice
+            adfc = doc.trade.agreement.buyer
+            user = doc.trade.agreement.seller
+
+        # Seller information
+        adfc.name = CONTACT['Name']
+        adfc.address.line_one = CONTACT['LineOne']
         if 'LineTwo' in CONTACT.keys():
-            doc.trade.agreement.seller.address.line_two = CONTACT['LineTwo']
+            adfc.address.line_two = CONTACT['LineTwo']
         if 'LineThree' in CONTACT.keys():
-            doc.trade.agreement.seller.address.line_three = CONTACT['LineThree']
-        doc.trade.agreement.seller.address.postcode = CONTACT['PostCode']
-        doc.trade.agreement.seller.address.city_name = CONTACT['City']
-        #doc.trade.agreement.seller.address.country_subdivision = CONTACT['State']
-        doc.trade.agreement.seller.address.country_id = CONTACT['Country']
+            adfc.address.line_three = CONTACT['LineThree']
+        adfc.address.postcode = CONTACT['PostCode']
+        adfc.address.city_name = CONTACT['City']
+        #adfc.address.country_subdivision = CONTACT['State']
+        adfc.address.country_id = CONTACT['Country']
 
         tr = DH_TaxRegistration()
-        tr.id = (CONTACT['VAT-Nr.'],"VA")
-        doc.trade.agreement.seller.tax_registrations.add(tr)
+        tr.id = ("VA",CONTACT['VAT-Nr.']) # FC == Tax number, VA == VAT Number
+        adfc.tax_registrations.add(tr)
+        user.tax_registrations.add(tr)
 
-        doc.trade.agreement.buyer.name = self.getusername()
-        doc.trade.agreement.buyer.description = self.getusergroup()
+        # Buyer information
+        user.name = self.getusername()
+        user.description = self.getusergroup()
+        user.address.line_one = "c/o " + CONTACT['Name']
+        user.address.line_two = CONTACT['LineOne']
+        if 'LineTwo' in CONTACT.keys():
+            user.address.line_three = CONTACT['LineTwo']
+        user.address.postcode = CONTACT['PostCode']
+        user.address.city_name = CONTACT['City']
+        #user.address.country_subdivision = CONTACT['State']
+        user.address.country_id = CONTACT['Country']
 
+        # Positions
         for index in range(self._POSITIONCOUNT):
             position = self.positions[index]
             if not position:
                 continue
             li = DH_LineItem()
-            li.document.line_id = str(index)
+            li.document.line_id = str(index+1)
             li.product.name = position.getname()
             li.agreement.net.amount = Decimal(abs(position.getunitprice()) or abs(position.getvalue()))
-            count = Decimal(position.getunitcount())
+            count = Decimal(position.getunitcount() * mode)
             if position.getcost():
                 count = -count
             li.agreement.net.basis_quantity = (count, "H87")  # H87 == Item
             li.delivery.billed_quantity = (count, "H87")  # H87 == Item
-            li.settlement.trade_tax.category_code = "K" # VAT exempt for intra community supply of goods
-            li.settlement.monetary_summation.total_amount = Decimal(position.getvalue())
+            li.settlement.trade_tax.type_code = "VAT"
+            li.settlement.trade_tax.category_code = 'E' # Exempt from tax
+            li.settlement.trade_tax.rate_applicable_percent = Decimal("0.00")
+            li.settlement.monetary_summation.total_amount = Decimal(position.getvalue() * mode)
             doc.trade.items.add(li)
 
         if self.getdonations():
             li = DH_LineItem()
-            li.document.line_id = "DONATIONS"
-            li.product.name = "Donations"
-            li.settlement.trade_tax.category_code = "K" # VAT exempt for intra community supply of goods
-            li.settlement.monetary_summation.total_amount = Decimal(self.getdonations())
+            li.document.line_id = "DONATION"
+            li.product.name = "Spenden"
+            li.agreement.net.amount = Decimal(self.getdonations())
+            li.agreement.net.basis_quantity = (Decimal(mode), "H87")  # H87 == Item
+            li.delivery.billed_quantity = (Decimal(mode), "H87")  # H87 == Item
+            li.settlement.trade_tax.type_code = "VAT"
+            li.settlement.trade_tax.category_code = 'E' # Exempt from tax
+            li.settlement.trade_tax.rate_applicable_percent = Decimal("0.00")
+            li.settlement.monetary_summation.total_amount = Decimal(self.getdonations() * mode)
             doc.trade.items.add(li)
+        
+        # Payment information
+        if mode == 1:
+            doc.trade.settlement.payment_means.payee_account.account_name = CONTACT['AccName']
+            doc.trade.settlement.payment_means.payee_account.iban = CONTACT['IBAN']
+            doc.trade.settlement.payment_means.payee_institution.bic = CONTACT['BIC']
+        else:
+            doc.trade.settlement.payment_means.payer_account.iban = CONTACT['IBAN']
+        
+        term = DH_PaymentTerms()
 
-        """
-        if self.getibanmode() == 1 and self.gettotal < 0:
+        if self.getibanmode() == 1 and self.gettotal() < 0:
             # Refund via bank transfer
-            pass
-        elif self.getibanmode() == 2 and self.gettotal > 0:
+            doc.trade.settlement.payment_means.type_code = "42" # Payment to bank account
+            term.description = "Wir überweisen den Abrechnungsbetrag auf dein Konto."
             if self.getibanknown():
-                pass
+                doc.trade.settlement.payment_means.information.add("Meine Bankverbindung ist dem ADFC Hamburg bekannt.")
             else:
-                doc.trade.settlement.payment_means.type_code = "59" # SEPA direct debit
-        elif self.getibanmode() == 3 and self.gettotal > 0:
+                doc.trade.settlement.payment_means.payee_account.iban = self.getaccountiban()
+                if self.getaccountname():
+                    doc.trade.settlement.payment_means.payee_account.account_name = self.getaccountname()
+        elif self.getibanmode() == 2 and self.gettotal() > 0:
+            doc.trade.settlement.payment_means.type_code = "59" # SEPA direct debit
+            term.description = "Wir ziehen den Abrechnungsbetrag per SEPA-Lastschrift ein."
+            if self.getsepamode() in (2,3):
+                if self.getsepamode() == 2:
+                    doc.trade.settlement.payment_means.information.add("Ein SEPA-Mandat liegt noch nicht vor.")
+                else:
+                    doc.trade.settlement.payment_means.information.add("Das vorliegende SEPA-Mandat ist veraltet.")
+                note = DH_IncludedNote()
+                note.content_code = "SPF"
+                note.content.add("Bitte senden Sie mir ein SEPA-Mandatsformular zu.")
+                note.subject_code = "AAI" # General information
+                doc.header.notes.add(note)
+        elif self.getibanmode() == 3 and self.gettotal() > 0:
+            term.description = "Du überweist den Abrechnungsbetrag selbst."
             doc.trade.settlement.payment_means.type_code = "42" # Payment to bank account
         else:
             # Fallback if none of the above applies
             doc.trade.settlement.payment_means.type_code = "ZZZ" # Mutually defined
-        """
+        
+        doc.trade.settlement.terms.add(term)
 
-        doc.trade.settlement.payment_means.type_code = "ZZZ" # Mutually defined
+        # Tax
+        total = Decimal(abs(self.gettotal()))
 
-        tax = DH_ApplicableTradeTax()
-        tax.category_code = "K" # VAT exempt for intra community supply of goods
-        doc.trade.settlement.trade_tax.add(tax)
+        trade_tax = DH_ApplicableTradeTax()
+        trade_tax.calculated_amount = Decimal("0.00")
+        trade_tax.basis_amount = total
+        trade_tax.type_code = "VAT"
+        trade_tax.category_code = 'E' # Exempt from tax
+        trade_tax.exemption_reason = "Vereinsinterne Abrechnung"
+        trade_tax.rate_applicable_percent = Decimal("0.00")
+        doc.trade.settlement.trade_tax.add(trade_tax)
 
+        # Total
         doc.trade.settlement.currency_code = "EUR"
-        doc.trade.settlement.monetary_summation.line_total = Decimal(self.gettotal())
-        doc.trade.settlement.monetary_summation.charge_total = Decimal(
-            sum(position.getincome() for position in self.positions)
-            + self.getdonations() )
-        doc.trade.settlement.monetary_summation.allowance_total = Decimal(
-            sum(position.getcost() for position in self.positions) )
-        doc.trade.settlement.monetary_summation.tax_basis_total = Decimal(self.gettotal())
-        doc.trade.settlement.monetary_summation.tax_total = Decimal("0.00")
-        doc.trade.settlement.monetary_summation.grand_total = Decimal(self.gettotal())
-        doc.trade.settlement.monetary_summation.due_amount = Decimal(self.gettotal())
+        doc.trade.settlement.monetary_summation.line_total = total
+        doc.trade.settlement.monetary_summation.charge_total = Decimal("0.00")
+        doc.trade.settlement.monetary_summation.allowance_total = Decimal("0.00")
+        doc.trade.settlement.monetary_summation.tax_basis_total = total
+        doc.trade.settlement.monetary_summation.tax_total = (Decimal("0.00"),"EUR")
+        doc.trade.settlement.monetary_summation.grand_total = total
+        doc.trade.settlement.monetary_summation.due_amount = total
 
-        # Return test file (for now)
-        with open('static/test.xml','rb') as file:
-            xml = file.read()
-        return xml
-
-        #return doc.serialize(schema="FACTUR-X_EXTENDED")
+        return doc.serialize(schema="FACTUR-X_EXTENDED")
     
     # Variable getters and setters
     def setusername(self,value:str = ""):